+++ sqlite3/lib/src/ffi/bindings.dart
@@ -913,6 +913,11 @@ final class FfiValue extends RawSqliteValue {
   int sqlite3_value_type() {
     return bindings.sqlite3_value_type(value);
   }
+
+  @override
+  int sqlite3_value_subtype() {
+    return bindings.sqlite3_value_subtype(value);
+  }
 }
 
 final class FfiContext extends RawSqliteContext {