+++ sqlite3/lib/src/implementation/utils.dart
@@ -9,13 +9,17 @@ extension BigIntRangeCheck on BigInt {
   }
 }
 
-int eTextRep(bool deterministic, bool directOnly) {
+int eTextRep(bool deterministic, bool directOnly, bool subtype) {
   var flags = SqlTextEncoding.SQLITE_UTF8;
   if (deterministic) {
     flags |= SqlFunctionFlag.SQLITE_DETERMINISTIC;
   }
   if (directOnly) {
     flags |= SqlFunctionFlag.SQLITE_DIRECTONLY;
+  }
+  if (subtype) {
+    flags |=
+        SqlFunctionFlag.SQLITE_SUBTYPE | SqlFunctionFlag.SQLITE_RESULT_SUBTYPE;
   }
 
   return flags;