+++ sqlite3/test/wasm/sqlite3_test.dart
@@ -45,7 +45,7 @@ void main() {
         expect(
           version,
           isA<Version>()
-              .having((e) => e.libVersion, 'libVersion', startsWith('3.48')),
+              .having((e) => e.libVersion, 'libVersion', startsWith('3.50')),
         );
       });